+++ qr_code/tests/tests.py
@@ -1,5 +1,6 @@
 """Tests for qr_code application."""
 import os
+from decimal import Decimal
 
 from django.test import SimpleTestCase
 from pydantic import ValidationError
@@ -46,7 +47,10 @@ class TestQRCodeOptions(SimpleTestCase):
 
     def test_kw_save(self):
         options = QRCodeOptions(border=0, image_format="png", size=13)
-        self.assertDictEqual(options.kw_save(), {"border": 0, "kind": "png", "scale": 13})
+        self.assertDictEqual(options.kw_save(), {'border': 0, 'dark': '#000', 'kind': 'png', 'light': '#fff', 'scale': 13})
+        options = QRCodeOptions(border=0, image_format="svg", size=13)
+        self.assertDictEqual(options.kw_save(),
+                             {'border': 0, 'dark': '#000', 'kind': 'svg', 'light': '#fff', 'scale': Decimal("1.3"), 'unit': 'mm'})
 
 
 class TestWriteResourceData(SimpleTestCase):