+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 # 4.0.0 (2024-01-03)
 * Upgrade dependencies, and drop support for Pydantic <2.5, Django<4.2
-* Add support for Python 3.12
+* Get rid of pytz dependency.
+* Add support for Python 3.12.
 * Add support for Django 5.0.
 * Introduce support for floating point size QR code modules.
-* Introduce embedded Base64 SVG image.
+* Introduce embedded Base64 SVG image as data URI.
 * Behavior change: white background is no longer interpreted as transparent for SVG output (#41). If you want to avoid path fill of SVG to reduce the size of the SVG image, you have to explicitly set `light_color` to `None` in `QRCodeOptions`.
 
 ## 3.1.2 (2023-04-10)